uses node.js astro adapter
