+++ examples/astro/website/astro.config.mjs
@@ -1,5 +1,13 @@
 // @ts-check
 import { defineConfig } from 'astro/config'
+import node from '@astrojs/node'
+import { loadEnv } from 'payload/node'
 
-// https://astro.build/config
-export default defineConfig({})
+loadEnv()
+
+export default defineConfig({
+  output: 'server',
+  adapter: node({
+    mode: 'standalone',
+  }),
+})
