+++ examples/astro/website/src/pages/index.astro
@@ -1,9 +1,10 @@
 ---
+import { getPayload } from 'payload'
+import { config } from 'payload-app'
 import Welcome from '../components/Welcome.astro'
 import Layout from '../layouts/Layout.astro'
-import { getPayload } from '../utils/getPayload'
 
-const payload = await getPayload()
+const payload = await getPayload({ config })
 
 const posts = await payload.find({ collection: 'posts', sort: 'createdAt' })
 // Welcome to Astro! Wondering what to do next? Check out the Astro documentation at https://docs.astro.build
