--- examples/astro/website/src/utils/getPayload.ts
@@ -1,4 +1,0 @@
-import { getPayload as _getPayload } from 'payload'
-import { resolveConfig } from 'payload-app'
-
-export const getPayload = () => _getPayload({ config: resolveConfig(import.meta.env) })
